Made graphics and scenes folder
The Graphics folder will hold all the pixel sprites for the game. As well as well as the soundtrack for the game. It will contain sub folders like player and enemy. The scenes folder will hold the what will be either levels or just one main game. It really just contains the main node.